+++ packages/spacecat-shared-data-access/test/unit/service/key-events/index.test.js
@@ -12,7 +12,7 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
@@ -20,10 +20,8 @@ import sinonChai from 'sinon-chai';
 import { keyEventFunctions } from '../../../../src/service/key-events/index.js';
 import { KEY_EVENT_TYPES } from '../../../../src/models/key-event.js';
 
-chai.use(chaiAsPromised);
-chai.use(sinonChai);
-
-const { expect } = chai;
+use(chaiAsPromised);
+use(sinonChai);
 
 const TEST_DA_CONFIG = {
   tableNameKeyEvents: 'spacecat-services-key-events',