+++ packages/spacecat-shared-slack-client/test/clients/base-slack-client.test.js
@@ -12,16 +12,14 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import nock from 'nock';
 import sinon from 'sinon';
 
 import { BaseSlackClient, SLACK_TARGETS } from '../../src/index.js';
 
-chai.use(chaiAsPromised);
-
-const { expect } = chai;
+use(chaiAsPromised);
 
 describe('BaseSlackClient', () => {
   const mockToken = 'mock-token';
@@ -174,7 +172,7 @@ describe('BaseSlackClient', () => {
 
       try {
         await client.fileUpload(file);
-      } catch (e) {
+      } catch {
         expect(mockLog.error.called).to.be.true;
       }
     });