+++ packages/spacecat-shared-utils/src/s3.js
@@ -12,17 +12,26 @@
 
 import { S3Client } from '@aws-sdk/client-s3';
 
-export function s3Bucket(fn) {
-  return async (req, context) => {
+/**
+ * Adds an S3Client instance and bucket to the context.
+ *
+ * @param {UniversalAction} fn
+ * @returns {function(object, UniversalContext): Promise<Response>}
+ */
+export function s3Wrapper(fn) {
+  return async (request, context) => {
     if (!context.s3) {
+      context.s3 = {};
+
       const {
         AWS_REGION: region,
         S3_BUCKET_NAME: bucket,
       } = context.env;
 
-      context.s3 = new S3Client({ region });
-      context.s3Bucket = bucket;
+      context.s3.s3Client = new S3Client({ region });
+      context.s3.s3Bucket = bucket;
     }
-    return fn(req, context);
+
+    return fn(request, context);
   };
 }