+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-google-client-v1.1.19](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.18...@adobe/spacecat-shared-google-client-v1.1.19) (2024-08-19)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#332](https://github.com/adobe/spacecat-shared/issues/332)) ([b2bb2a4](https://github.com/adobe/spacecat-shared/commit/b2bb2a4dd1a0995c36cd2b64510b7c4e0e05436f))
+
+# [@adobe/spacecat-shared-google-client-v1.1.18](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.17...@adobe/spacecat-shared-google-client-v1.1.18) (2024-08-14)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#320](https://github.com/adobe/spacecat-shared/issues/320)) ([c75b743](https://github.com/adobe/spacecat-shared/commit/c75b7432e0add9b261ddc7999fe80b20442a0dd7))
+
 # [@adobe/spacecat-shared-google-client-v1.1.17](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.16...@adobe/spacecat-shared-google-client-v1.1.17) (2024-08-06)